+++ Control/Joint/Base/Maybe.hs
@@ -1,7 +1,7 @@
 module Control.Joint.Base.Maybe where
 
-import Control.Joint.Composition (Composition (Primary, unwrap))
-import Control.Joint.Transformer (Transformer (Schema, lay, wrap))
+import Control.Joint.Composition (Composition (Primary, run))
+import Control.Joint.Transformer (Transformer (Schema, embed, build))
 import Control.Joint.Schemes.UT (UT (UT))
 
 instance Functor u => Functor (UT Maybe u) where
@@ -12,13 +12,13 @@ instance Applicative u => Applicative (UT Maybe u) where
 	UT f <*> UT x = UT $ (<*>) <$> f <*> x
 
 instance (Applicative u, Monad u) => Monad (UT Maybe u) where
-	UT x >>= f = UT $ x >>= maybe (pure Nothing) (unwrap . f)
+	UT x >>= f = UT $ x >>= maybe (pure Nothing) (run . f)
 
 instance Composition Maybe where
 	type Primary Maybe a = Maybe a
-	unwrap x = x
+	run x = x
 
 instance Transformer Maybe where
 	type Schema Maybe u = UT Maybe u
-	lay x = UT $ Just <$> x
-	wrap x = UT . pure $ x
+	embed x = UT $ Just <$> x
+	build x = UT . pure $ x