+++ Control/Joint/Base/State.hs
@@ -1,8 +1,8 @@
 module Control.Joint.Base.State where
 
 import Control.Joint.Core (type (:.), type (:=))
-import Control.Joint.Composition (Composition (Primary, unwrap))
-import Control.Joint.Transformer (Transformer (Schema, lay, wrap))
+import Control.Joint.Composition (Composition (Primary, run))
+import Control.Joint.Transformer (Transformer (Schema, embed, build))
 import Control.Joint.Schemes.TUT (TUT (TUT))
 
 newtype State s a = State ((->) s :. (,) s := a)
@@ -25,12 +25,12 @@ instance Monad (State s) where
 
 instance Composition (State s) where
 	type Primary (State s) a = (->) s :. (,) s := a
-	unwrap (State x) = x
+	run (State x) = x
 
 instance Transformer (State s) where
 	type Schema (State s) u = TUT ((->) s) u ((,) s)
-	lay x = TUT $ \s -> (s,) <$> x
-	wrap x = TUT $ pure <$> unwrap x
+	embed x = TUT $ \s -> (s,) <$> x
+	build x = TUT $ pure <$> run x
 
 instance Functor u => Functor (TUT ((->) s) u ((,) s)) where
 	fmap f (TUT x) = TUT $ \old -> (fmap . fmap) f $ x old
@@ -40,7 +40,7 @@ instance Monad u => Applicative (TUT ((->) s) u ((,) s)) where
 	TUT f <*> TUT x = TUT $ \old -> f old >>= \(new, g) -> (fmap . fmap) g $ x new
 
 instance Monad u => Monad (TUT ((->) s) u ((,) s)) where
-	TUT x >>= f = TUT $ \old -> x old >>= \(new, y) -> ($ new) . unwrap . f $ y
+	TUT x >>= f = TUT $ \old -> x old >>= \(new, y) -> ($ new) . run . f $ y
 
 get :: State s s
 get = State $ \s -> (s, s)