+++ Control/Joint/Composition.hs
@@ -1,6 +1,6 @@
 module Control.Joint.Composition (Composition (..)) where
 
 class Composition t where
-	{-# MINIMAL unwrap #-}
+	{-# MINIMAL run #-}
 	type Primary t a :: *
-	unwrap :: t a -> Primary t a
+	run :: t a -> Primary t a