+++ Control/Joint/Transformer.hs
@@ -4,10 +4,10 @@ import Control.Joint.Core (type (~>))
 import Control.Joint.Composition (Composition)
 
 class Composition t => Transformer t where
-	{-# MINIMAL lay, wrap #-}
+	{-# MINIMAL embed, build #-}
 	type Schema (t :: * -> *) (u :: * -> *) = (r :: * -> *) | r -> t u
-	lay :: Functor u => u ~> Schema t u
-	wrap :: Applicative u => t ~> Schema t u
+	embed :: Functor u => u ~> Schema t u
+	build :: Applicative u => t ~> Schema t u
 
 infixr 1 :>
 type (:>) t u a = Transformer t => Schema t u a