+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 * Define `Functor`/`Applicative`/`Monad` instances for transformers
 * Define `State` datatype with `get`, `modify` and `put` operations
 * Rename and fix `UTU` joint scheme to `TUT`
+* Renaming: `unwrap` -> `run`, `wrap` -> `build`, `lay` -> `embed`